+++ Corgibytes.Freshli.Cli/CommandRunners/Cache/DestroyCommandRunner.cs
@@ -3,12 +3,11 @@ using System.CommandLine.Invocation;
 using System.CommandLine.IO;
 using Corgibytes.Freshli.Cli.CommandOptions;
 using Corgibytes.Freshli.Cli.Commands;
-using Corgibytes.Freshli.Cli.Commands.Cache;
 using Corgibytes.Freshli.Cli.Extensions;
+using Corgibytes.Freshli.Cli.Functionality.CacheDestroy;
 using Corgibytes.Freshli.Cli.Functionality.Engine;
 using Corgibytes.Freshli.Cli.Resources;
 using Corgibytes.Freshli.Lib;
-using Elasticsearch.Net;
 
 namespace Corgibytes.Freshli.Cli.CommandRunners.Cache;
 
@@ -29,11 +28,10 @@ public class CacheDestroyCommandRunner : CommandRunner<CacheCommand, CacheDestro
     {
         ActivityEngine.Dispatch(new DestroyCacheActivity(options, context));
 
-        bool isConfirmationRequired = false;
+        var isConfirmationRequired = false;
         EventEngine.On<ConfirmationRequiredEvent>(_ => { isConfirmationRequired = true; });
 
-        int exitCode = 0;
-        exitCode = WaitForCacheDestroyEvents(context);
+        var exitCode = WaitForCacheDestroyEvents(context);
         ActivityEngine.Wait();
 
         if (isConfirmationRequired)
@@ -58,7 +56,7 @@ public class CacheDestroyCommandRunner : CommandRunner<CacheCommand, CacheDestro
 
     private int WaitForCacheDestroyEvents(InvocationContext context)
     {
-        int exitCode = 0;
+        var exitCode = true.ToExitCode();
 
         EventEngine.On<CacheDestroyFailedEvent>(destroyEvent =>
         {
@@ -66,7 +64,10 @@ public class CacheDestroyCommandRunner : CommandRunner<CacheCommand, CacheDestro
             exitCode = false.ToExitCode();
         });
 
-        EventEngine.On<CacheDestroyedEvent>(_ => { });
+        EventEngine.On<CacheDestroyedEvent>(destroyEvent =>
+        {
+            exitCode = destroyEvent.ExitCode;
+        });
 
         ActivityEngine.Wait();
