+++ Corgibytes.Freshli.Cli/Functionality/CacheDestroy/CacheDestroyFailedEvent.cs
@@ -1,6 +1,6 @@
 using Corgibytes.Freshli.Cli.Functionality.Engine;
 
-namespace Corgibytes.Freshli.Cli.Commands.Cache;
+namespace Corgibytes.Freshli.Cli.Functionality.CacheDestroy;
 
 public class CacheDestroyFailedEvent : IApplicationEvent
 {
