+++ Corgibytes.Freshli.Cli/Functionality/CacheDestroy/DestoryCacheActivity.cs
@@ -2,32 +2,27 @@ using System.CommandLine.Invocation;
 using System.CommandLine.IO;
 using Corgibytes.Freshli.Cli.CommandOptions;
 using Corgibytes.Freshli.Cli.Extensions;
-using Corgibytes.Freshli.Cli.Functionality;
 using Corgibytes.Freshli.Cli.Functionality.Engine;
 using Corgibytes.Freshli.Cli.Resources;
 
-namespace Corgibytes.Freshli.Cli.Commands.Cache;
+namespace Corgibytes.Freshli.Cli.Functionality.CacheDestroy;
 
 public class DestroyCacheActivity : IApplicationActivity
 {
     public DestroyCacheActivity(CacheDestroyCommandOptions options, InvocationContext context)
     {
-        cmdOptions = options;
-        cmdContext = context;
+        CmdOptions = options;
+        CmdContext = context;
     }
 
-    private CacheDestroyCommandOptions cmdOptions { get; }
-    private InvocationContext cmdContext { get; }
+    private CacheDestroyCommandOptions CmdOptions { get; }
+    private InvocationContext CmdContext { get; }
 
     public void Handle(IApplicationEventEngine eventClient)
     {
-        var strConfirmDestroy = string.Format(
-            CliOutput.CacheDestroyCommandRunner_Run_Prompt,
-            cmdOptions.CacheDir.FullName);
-
         // Unless the --force flag is passed, prompt the user whether they want
         // to destroy the cache
-        if (!cmdOptions.Force)
+        if (!CmdOptions.Force)
         {
             eventClient.Fire(new ConfirmationRequiredEvent());
             return;
@@ -35,20 +30,18 @@ public class DestroyCacheActivity : IApplicationActivity
 
         var strDestroyingCache = string.Format(
             CliOutput.CacheDestroyCommandRunner_Run_Destroying,
-            cmdOptions.CacheDir);
+            CmdOptions.CacheDir);
 
         // Destroy the cache
-        cmdContext.Console.Out.WriteLine(strDestroyingCache);
+        CmdContext.Console.Out.WriteLine(strDestroyingCache);
         try
         {
-            Functionality.Cache.Destroy(cmdOptions.CacheDir).ToExitCode();
+            var exitCode = Cache.Destroy(CmdOptions.CacheDir).ToExitCode();
+            eventClient.Fire(new CacheDestroyedEvent() {ExitCode = exitCode});
         }
         catch (CacheException error)
         {
             eventClient.Fire(new CacheDestroyFailedEvent() {ResultMessage = error.Message});
-            return;
         }
-
-        eventClient.Fire(new CacheDestroyedEvent());
     }
 }
